Added tests to Binary vs Arma to compare matrix access and multiplication

+++ Binary_vs_Arma/Binary_vs_Arma/BinaryMatrix.h
@@ -11,7 +11,7 @@
 
 using namespace arma;
 
-typedef uint64_t uint8;
+typedef uint8_t uint8;
 typedef unsigned int uint;
 typedef std::pair<uint, uint> uIntPair;
